chore: favourite contacts

+++ apps/contacts/views.py
@@ -4,10 +4,10 @@ from rest_framework.response import Response
 from drf_spectacular.utils import extend_schema, OpenApiParameter, OpenApiExample
 from rest_framework.pagination import PageNumberPagination
 
-from .models import Contact
-from .serializers import ContactSerializer
+from .models import Contact, Favourite
+from .serializers import ContactSerializer, FavouritesSerializer
 
-tags = ["Contacts"]
+tags = [["Contacts"], ["Favourites"]]
 
 # V2
 # TODO favourites list
@@ -20,20 +20,22 @@ class ContactListCreateAPIView(APIView, PageNumberPagination):
     page_size = 10
 
     @extend_schema(
-        tags=tags,
+        tags=tags[0],
         summary="Retrieve contacts",
         description="This endpoint retrieves contacts belonging to the authenticated user.",
         responses={200: ContactSerializer},
         # parameters=[OpenApiParameter(name="search", description="Search term", required=False, type=str)]
     )
     def get(self, request):
-        contacts = Contact.objects.filter(owner=request.user).order_by("first_name", "last_name")
+        contacts = Contact.objects.filter(owner=request.user).order_by(
+            "first_name", "last_name"
+        )
         paginated_objects = self.paginate_queryset(contacts, request, view=self)
         serializer = self.serializer_class(paginated_objects, many=True)
         return self.get_paginated_response({"data": serializer.data})
 
     @extend_schema(
-        tags=tags,
+        tags=tags[0],
         summary="Create a new contact",
         description="This endpoint creates a new contact for the authenticated user.",
         request=ContactSerializer,
@@ -56,21 +58,21 @@ class ContactListCreateAPIView(APIView, PageNumberPagination):
         serializer = self.serializer_class(data=request.data)
         serializer.is_valid(raise_exception=True)
         serializer.save(owner=request.user)
-        #remove the created_at and updated_at fields from the response
+        # remove the created_at and updated_at fields from the response
         data = serializer.data
         data.pop("created_at")
         data.pop("updated_at")
         return Response({"data": data}, status=201)
 
 
-class ContactDetailView(APIView):
+class ContactDetailAPIView(APIView):
     permission_classes = [IsAuthenticated]
     serializer_class = ContactSerializer
 
     @extend_schema(
-        tags=tags,
-        summary="Retrieve a  contact's detail",
-        description="This endpoint a contact's detail belonging to the authenticated user.",
+        tags=tags[0],
+        summary="Retrieve a contact's detail",
+        description="This endpoint retreives a contact's detail",
         responses={200: ContactSerializer},
     )
     def get(self, request, pk):
@@ -83,9 +85,9 @@ class ContactDetailView(APIView):
             return Response({"error": "Contact not found"}, status=404)
 
     @extend_schema(
-        tags=tags,
-        summary="Update a  contact's detail",
-        description="This endpoint updates a contact's detail belonging to the authenticated user.",
+        tags=tags[0],
+        summary="Update a contact's detail",
+        description="This endpoint updates a contact's detail",
         responses={200: ContactSerializer},
     )
     def put(self, request, pk):
@@ -96,12 +98,71 @@ class ContactDetailView(APIView):
             return Response(serializer.data)
 
     @extend_schema(
-        tags=tags,
+        tags=tags[0],
         summary="Delete a contact",
         description="This endpoint deletes a contact belonging to the authenticated user.",
         responses={204: ContactSerializer},
     )
-    def delete(self, request, pk):
-        contact = Contact.objects.get(pk=pk)
-        contact.delete()
-        return Response({"Successsful": "Deleted successfully"})
+    def delete(self, request, id):
+        contact = Contact.objects.filter(id=id).first()
+        if contact:
+            contact.delete()
+            return Response({"Successsful": "Deleted successfully"})
+        else:
+            return Response({"error": "Contact not found"}, status=404)
+
+
+class FavouritesListCreateAPIView(APIView):
+    serializer_class = FavouritesSerializer
+    permission_classes = [IsAuthenticated]
+
+    @extend_schema(
+        tags=tags[1],
+        summary="Retrieve favourites",
+        description="This endpoint retreives the authenticated user's favourite contacts.",
+        responses={200: FavouritesSerializer},
+    )
+    def get(self, request):
+        favourites = Favourite.objects.filter(owner=request.user)
+        serializer = self.serializer_class(favourites, many=True)
+        return Response({"data": serializer.data})
+
+    @extend_schema(
+        tags=tags[1],
+        summary="Create favourites",
+        description="This endpoint adds a new contact to the authenticated user's favourite contacts.",
+        request=FavouritesSerializer,
+        responses={201: FavouritesSerializer},
+    )
+    def post(self, request):
+        serializer = self.serializer_class(data=request.data)
+        serializer.is_valid(raise_exception=True)
+        contact = serializer.validated_data["contact"]
+        contact.is_favourite = True
+        contact.save()
+        serializer.save(owner=request.user)
+        return Response({"data": serializer.data}, status=201)
+
+
+class FavouriteDetailAPIView(APIView):
+    serializer_class = FavouritesSerializer
+    permission_classes = [IsAuthenticated]
+
+    @extend_schema(
+        tags=tags[1],
+        summary="Remove favourite",
+        description="This endpoint removes a contact from user's favourites list",
+        responses={204: ContactSerializer},
+    )
+    def delete(self, request, id):
+        favourite = Favourite.objects.filter(id=id).first()
+        
+        if favourite:
+            contact = favourite.contact
+            
+            contact.is_favourite = False
+            contact.save()
+            favourite.delete()
+            return Response({"Successsful": "Deleted successfully"})
+        else:
+            return Response({"error": "Contact not found"})
